Se separo el armado de la descripcion de la busqueda externa
